% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{abs.proj}
\alias{abs.proj}
\title{Internal - Rescales numeric matrix for projection}
\usage{
\method{abs}{proj}(proj, abs, quant.low, quant.high)
}
\arguments{
\item{proj}{a data.frame providing the manifold representation}

\item{abs}{a boolean value providing xx}

\item{quant.low}{a numeric value providing the number first quantile}

\item{quant.high}{a numeric value providing the number last quantile}
}
\value{
a numeric vector containing scale limits
}
\description{
This functions is used internally xxx
}
