% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Celldata-class}
\alias{Celldata-class}
\alias{Celldata}
\title{Celldata class definition}
\description{
The Celldata object is a S4 object containing all single-cell information.
}
\section{Slots}{

\describe{
\item{\code{samples}}{a character vector containing the names of the biological samples}

\item{\code{raw.markers}}{a character vector containing the names of the raw markers}

\item{\code{matrix.expression.r}}{a data.frame containing the raw marker expressions of each cell}

\item{\code{matrix.expression}}{a data.frame containing the marker expressions of each cell}

\item{\code{manifold}}{a data.frame containing the manifold coordinates}

\item{\code{manifold.params}}{a list containing the parameters used for manifold generation}

\item{\code{identify.clusters}}{a vector containing the names of the identified cell clusters}

\item{\code{identify.clusters.params}}{a vector containing the parameters used for the identification of the cell clusters}

\item{\code{concave.hulls}}{a data.frame containing the coordinates of the concave hulls of each cluster}

\item{\code{matrix.cell.count}}{a data.frame containing the number of cells associated to each cluster for each sample}

\item{\code{matrix.abundance}}{a data.frame containing the percentage of cells associated to each cluster for each sample}

\item{\code{statistic}}{a data.frame containing the statistics of cell clusters}

\item{\code{metadata}}{a data.frame containing the metadata associated to each sample}
}}

