% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{computeCellCounts}
\alias{computeCellCounts}
\title{Internal - Computes the number of cells for each cluster}
\usage{
computeCellCounts(proj, clusters, samples)
}
\arguments{
\item{proj}{a data.frame providing the manifold representation with two columns}

\item{clusters}{a character vector providing the cluster to analyse the associated with cell cluster}

\item{samples}{a character vector providing for each cell the associated biological sample}
}
\value{
a data.frame containing the numbers of cells associated for each cluster for each sample (rownames = clusters / colnames = samples)
}
\description{
This function is used internally to computes the number of cells for each cluster.
}
