% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{computeCellDensities}
\alias{computeCellDensities}
\title{Internal - Computes the cell density within the manifold}
\usage{
computeCellDensities(proj, n)
}
\arguments{
\item{proj}{a data.frame providing the manifold representation with two columns}

\item{n}{a numerical providing the number a grid points in each direction (please refer to the function 'kde2d' of the 'MASS' package)}
}
\value{
a numeric vector containing the computed cell density
}
\description{
This function is used internally to compute the cell density
}
