% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{computeClusterAbundances}
\alias{computeClusterAbundances}
\title{Internal - Computes the abundances for each cell cluster}
\usage{
computeClusterAbundances(count)
}
\arguments{
\item{count}{a data.frame providing the numbers of cells associated to each cluster for each sample}
}
\value{
a data.frame containing the abundance of cells to each clusters for each sample
}
\description{
This function is used internally to computes the abundance of each cluster for each sample.
}
