% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{computeConcaveHulls}
\alias{computeConcaveHulls}
\title{Internal - Computes the concave hulls for identified cell clusters}
\usage{
computeConcaveHulls(proj, clusters, concavity = 2, length.threshold = 0)
}
\arguments{
\item{proj}{a data.frame providing the manifold representation}

\item{clusters}{a character vector providing id of cell clusters for which the concave hulls must be computed}

\item{concavity}{a numeric value providing a relative measure of concavity (please refer to the function 'concaveman' of the 'concaveman' package)}

\item{length.threshold}{a numeric value providing a threshold for the segment length (please refer to the function 'concaveman' of the concaveman package)}
}
\value{
a data.frame containing the concave hulls for each cluster (dim1, dim2, clusters)
}
\description{
This function is used internally to computes the concave hulls each cell cluster.
Each concave hulls correspond the boundaries of the cells cluster with manifold representation.
}
