% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{computeMarkerMedians}
\alias{computeMarkerMedians}
\title{Internal - Computes marker expression medians for each cluster}
\usage{
computeMarkerMedians(exprs)
}
\arguments{
\item{exprs}{a data.frame containing the marker expressions for each cell}
}
\value{
a data.frame containing the median of marker expressions for each cluster (clusters, markers, medians)
}
\description{
This function is used internally to computes marker expression medians for each cluster.
}
