% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeSmallClusters}
\alias{computeSmallClusters}
\title{Internal - Computes the percentage of clusters with low number of cells}
\usage{
computeSmallClusters(UMAPdata, th.size)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{th.size}{a numeric value providing the minimum number of cells needed for a cluster to be considered a small cluster}
}
\value{
a list containing QC information for small clusters. 
Returns a data.frame with a boolean value indicating if the number of associated cells is greater or less than the threshold.
In addition, returns the percentage of the calculation.
}
\description{
This function is used internally to compute the percentage of clusters having a number of associated cells lower than a specific threshold.
}
