% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Computes differential analysis statistics for cell clusters}
\usage{
computeStatistics(
  Celldata,
  condition,
  ref.condition,
  comparison = "cmp",
  test.statistics = c("wilcox.test", "t.test"),
  p.adjust = c("none", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"),
  paired = FALSE
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{condition}{a character value providing the name of the condition to be compared}

\item{ref.condition}{a character value providing the name of reference condition}

\item{comparison}{a character value providing the name of comparison}

\item{test.statistics}{a character value providing the type of statistical test to use. Possible values are: 'wilcoxon' or 't-test'}

\item{p.adjust}{a character value providing the type of p-value adjustment method to use. Possible values are: 'none', 'holm', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY', 'fdr'}

\item{paired}{a boolean value indicating if individuals are paired}
}
\value{
a S4 object of class 'Celldata'
}
\description{
This function aims to identify of differentially abundant clusters.

Such clusters correspond to cell clusters having abundances statistically different between two biological conditions.
The statistical test used for the comparisons can be defined by users.
For each cluster, the p-value, log2 fold-change and effect size relative to the reference condition are computed.
Statistical comparison can be performed in a paired and unpaired manner.
Computed p-values can be corrected for multiple testing.
}
