% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{computeUniformClusters}
\alias{computeUniformClusters}
\title{Internal - Computes percentage of clusters with uniform phenotype}
\usage{
computeUniformClusters(UMAPdata, uniform.test, th.pvalue, th.IQR)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{uniform.test}{a character providing the name of test assessment to perform. Possible value are : 'both', 'uniform', 'IQR'}

\item{th.pvalue}{a numeric value providing the p-value threshold of the Hartigan's dip test (unimodal if pvalue > th.pvalue)}

\item{th.IQR}{a numeric value providing the IQR (interquartile range) threshold to assume a distribution as uniform}
}
\value{
a list containing QC information for small clusters. 
Returns a data.frame with a boolean value indicating xxx.
In addition, returns the percentage of the calculation.
}
\description{
This function is used internally to identify cell clusters that have a non-uniform phenotype.
A uniform cluster corresponds to clusters that have a unimodal expression and having a low spread of expression for all the markers to compose it.
}
