% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{createFlowframe}
\alias{createFlowframe}
\title{Internal - Creates of a FlowFrame object}
\usage{
createFlowframe(intensities)
}
\arguments{
\item{intensities}{a data.frame providing the cell profile intensities}
}
\value{
a flowframe object containing the marker expression
}
\description{
This function is used internally to create a flowframe object with the purpose extracting marker intensities to FCS files.
}
