% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{deleteClusters}
\alias{deleteClusters}
\title{Delete cluster}
\usage{
deleteClusters(Celldata, clusters)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifiers of the clusters to delete}
}
\value{
a Celldata object
}
\description{
This function aims to delete a set of cell clusters from this analysis
}
