% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{export,Celldata-method}
\title{Exports cell expression profiles to TSV or FCS files}
\usage{
export(Celldata, filename, clusters = NULL, samples = NULL)

\S4method{export}{Celldata}(Celldata, filename, clusters = NULL, samples = NULL)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{filename}{a character value providing the name of the output file}

\item{clusters}{a character vector containing the identifiers of the cell clusters to export. By default, all clusters are extracted}

\item{samples}{a character vector containing the names of biological samples to export. By default, all samples are extracted}
}
\value{
none
}
\description{
Exports cell expresion profiles from a Celldata object to a tab-separated or FCS files

Cell expression profiles can be exported for a set of given samples and for a set of given cell clusters
}
