% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{generateManifold}
\alias{generateManifold}
\title{Generates a manifold of cell events}
\usage{
generateManifold(
  UMAPdata,
  type = c("UMAP", "tSNE", "lvish"),
  markers = NULL,
  seed = 42,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{type}{a character value specifying the type of manifold to compute. Possible values are: 'UMAP' for Uniform Manifold Approximation and Projection, 'tSNE' for t Stochastic Neighbor Embedding, and 'lvish' for LargeVis}

\item{markers}{a character vector providing the cell markers to use for the manifold generation}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}

\item{verbose}{a boolean value indicating if computational details must be displayed on the console}

\item{...}{Other arguments passed on to methods}
}
\value{
a S4 object of class 'UMAPdata'
}
\description{
This function aims to generate a manifold representation for cell events stored in a UMAPdata object

This function allows the use of several non-linear dimension reduction techniques such as UMAP, t-SNE or LargeVis
The whole set of cell markers or specific cell markers can be used during the dimensionality reduction process
}
