% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{generateManifoldtSNE}
\alias{generateManifoldtSNE}
\title{Internal - Generates a t-SNE manifold of cell events}
\usage{
generateManifoldtSNE(exprs, seed, verbose, ...)
}
\arguments{
\item{exprs}{a data.frame containing the marker expressions}

\item{seed}{a numeric value providing the random seed to use in stochastic operation}

\item{verbose}{a boolean value indicating if computational details must be displayed on the console (please refer to the function 'Rtsne' of the 'Rtsne' package)}

\item{...}{Other arguments passed on to methods}
}
\value{
a data.frame containing the manifold coordinates
}
\description{
This function is used internally to compute a manifold representation based on the t-SNE algorithms.
}
