% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{generateManifoldUMAP}
\alias{generateManifoldUMAP}
\title{Internal - Generates a UMAP manifold of cell events}
\usage{
generateManifoldUMAP(exprs, seed, verbose, ...)
}
\arguments{
\item{exprs}{a data.frame providing the marker expressions}

\item{seed}{a numeric value providing the random seed to use in stochastic operation}

\item{verbose}{a boolean value indicating if computational details must be displayed on the console}

\item{...}{Other arguments passed on to methods}
}
\value{
a data.frame containing the manifold coordinates
}
\description{
This function is used internally a manifold representation based on the UMAP algorithms.
}
