% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getSamples}
\alias{getSamples}
\title{Filter to select samples}
\usage{
getSamples(Celldata, individual = NULL, condition = NULL, timepoint = NULL)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{individual}{a character vector containing the names of biological individual to use}

\item{condition}{a character vector containing the biological condition to use}

\item{timepoint}{a character vector containing the timepoint condition to use}
}
\value{
a ggplot2 object
}
\description{
This function aims to select the samples of interest
}
