% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{identifyClusters}
\alias{identifyClusters}
\title{Identify cell cluster of having similar marker expressions}
\usage{
identifyClusters(
  UMAPdata,
  space = c("manifold", "markers"),
  markers = NULL,
  method = c("kmeans", "kmedian", "clara", "DBSCAN", "SOM"),
  concavity = 2,
  length.threshold = 0,
  seed = 42,
  ...
)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{space}{a character value containing the space of clustering method to use. Possible values are: 'manifold' or 'markers'}

\item{markers}{a character vector providing the cell markers to use for the manifold generation}

\item{method}{a character value containing the name of the clustering method to use. Possible values are: 'kmeans', 'kmedian', 'clara', 'DBSCAN' and 'SOM'}

\item{concavity}{a numeric value providing a relative measure of concavity for the computation of the concave hulls (please refer to the function 'concaveman' of the 'concaveman' package)}

\item{length.threshold}{a numeric value providing a threshold of the segment length for the computation of the concave hulls (please refer to the function 'concaveman' of the concaveman package)}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}

\item{...}{Other arguments passed on to methods}
}
\value{
a S4 object of class 'UMAPdata'
}
\description{
This function aims to identify cell clusters, which are groups of cells having similar expressions for selected markers, using different unsupervised clustering methods.

Several clustering method are available such as kmeans, kmedian, clara, DBSCAN, HDBSCAN and SOM.
The cell clustering can be performed on the manifold representation or based on marker expression.
}
\details{
For each identify cell cluster, the boundaries of cells belonging to this cluster are delineated using a concave hull
}
