% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Imports of cell expression profiles from TSV or FCS files}
\usage{
import(
  files,
  filetype = "fcs",
  transform = c("logicle", "arcsinh", "logarithmic", "none"),
  downsampling = NULL,
  d.method = c("uniform", "density"),
  exclude.markers = NULL,
  seed = 42
)
}
\arguments{
\item{files}{a character vector specifying the path of the tab-separated or FCS files to load}

\item{filetype}{a character vector specifying the format of the loaded files. By default, FCS is used}

\item{transform}{a character value containing the type of the transformation to apply. Possible values are: 'logicle', 'arcsinh', 'logarithmic' or 'none'}

\item{downsampling}{a numeric value providing the number of cells to downsample for each sample}

\item{d.method}{a character value containing the type of the downsampling to apply. Possible values are: 'uniform' or 'density'}

\item{exclude.markers}{a character vector providing the marker names to be excluded during the import}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}
}
\value{
a S4 object of class 'UMAPdata'
}
\description{
This function aims to import acquired cell events into a UMAPdata object.

Input files can be tab-separated or FCS files.
Different transformations can be applied such as logicle, arcsinh or logarithmic.
Importantly, a downsampling of cell events can be performed using uniformly-based or density-based random selections.
Cell marker having technical or biological biaises can be excluded during the import.
}
