% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importMTX}
\alias{importMTX}
\title{Imports of cell expression profiles from MTX files}
\usage{
importMTX(mtx = mtx, cells = cells, features = features)
}
\arguments{
\item{mtx}{a character vector specifying the path of the mtx file to load}

\item{cells}{a character vector specifying the format of the mtx file to load}

\item{features}{a character vector specifying the format of the mtx file to load}
}
\value{
a S4 object of class 'Celldata'
}
\description{
This function aims to import acquired cell events from single-cell transcriptomic profiling into a Celldata object.
}
