% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsampling.R
\name{performUpsampling}
\alias{performUpsampling}
\title{Performs the upsampling of downsampled events}
\usage{
performUpsampling(UMAPdata, files)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{files}{a character vector providing the path of the tab-separated or FCS files}
}
\value{
a S4 object of class 'UMAPdata'
}
\description{
This function aims to perform upsample downsampled events based on an existing UMAPdata object and existing cell events stored in tab-separated or FCS files. 

Importantly, the identification of cell clusters must have been performed prior to this operation.
}
