% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,UMAPdata,ANY-method}
\title{Plots graphics for all UMAPdata objects}
\usage{
\S4method{plot}{UMAPdata,ANY}(x)
}
\arguments{
\item{x}{a UMAPdata object}
}
\value{
a ggplot2 object
}
\description{
Generates a graphical representation for a UMAPdata object.
The displayed representation depends on the current analysis status of the UMAPdata object. 

- If the manifold has not been calculated, then the number of cells per sample will be displayed.

- If the manifold has been calculated but not the clustering, then the manifold representation will be displayed. 

- If the manifold and clustering have been calculated, then a heatmap of marker expressions will be displayed.
}
