% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Plots cell cluster abundances using a boxplot representation}
\usage{
plotBoxplot(
  UMAPdata,
  clusters,
  samples = NULL,
  observation = c("individual", "condition", "timepoint"),
  test.statistics = c("wilcox.test", "t.test"),
  paired = FALSE,
  hide.ns = TRUE
)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{clusters}{a character vector containing the identifiers of the clusters to use}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{observation}{a character value containing the parameters to use}

\item{test.statistics}{a character value providing the type of statistical test to use. Possible values are: 'wilcox.test' or 't.test'}

\item{paired}{a boolean value indicating if a paired or unpaired comparison should be applied}

\item{hide.ns}{a boolean value indicating xxx}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize and compare the cell cluster abundances for each biological condition using boxplot and jitter representations.

The abundance of a specific cell cluster or a set of cell clusters can be displayed. 
The representation can be resticted to a specific set of samples. 
Moreover boxplot can be constructed based on sample meta information. 
Statistic can be computed for all comparisons.
}
