% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotCellCounts}
\alias{plotCellCounts}
\title{Plots the numbers of cells for each sample}
\usage{
plotCellCounts(
  Celldata,
  stats = c("min", "median", "mean", "q75", "max"),
  samples = NULL,
  sort = TRUE
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{stats}{a character vector providing the statistics to display. Possible values are: 'min', 'median', 'mean', 'q75', 'max'}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{sort}{a boolean value indicating if clusters must be sorted by the number associated sample}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the number of cells associated to each sample.

This representation displays the samples in the X-axis and the number of associated cells in the Y-axis.
Several statistics can be computed and shown.
}
\details{
The following statistic can be computed:

-'min' corresponds to the lowest number of cells within a data set

-'median' corresponds to the number of cells separates the lower half from the upper half within data set

-'mean' corresponds to the number of cells quantity shared within data set

-'q75' corresponds to the number of cells separates the quantiles 75% within data set

-'max' corresponds to the largest number of cells within a data set
}
