% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotClustersCounts}
\alias{plotClustersCounts}
\title{Plots the numbers of cells for each clusters}
\usage{
plotClustersCounts(UMAPdata, clusters = NULL, sort = TRUE)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{sort}{a boolean value indicating if clusters must be sorted by the number associated cluster}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the number of cells associated to each clusters.

This representation displays the clusters in the X-axis and the total number of associated cells in the Y-axis.
}
