% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{plotCombineHM}
\alias{plotCombineHM}
\title{Plots a combined expression and statistic heatmaps}
\usage{
plotCombineHM(HM1, HM2)
}
\arguments{
\item{HM1}{a ggplot object containing the expression heatmap}

\item{HM2}{a ggplot object containing the statistic heatmap}
}
\value{
a ggplot2 object
}
\description{
This function aims to combine the expression and statistic heatmaps.
}
