% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotCompareClusters}
\alias{plotCompareClusters}
\title{Plot compare cluster}
\usage{
plotCompareClusters(Celldata, clusters1, clusters2)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters1}{a character vector containing the identifier of the cluster to use}

\item{clusters2}{a character vector containing the identifier of the cluster to use}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the expression of a given marker in each cluster.
}
