% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotCoordinates}
\alias{plotCoordinates}
\title{Plots parallel coordinates}
\usage{
plotCoordinates(UMAPdata)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}
}
\value{
xx
}
\description{
This function aims to visualize xxx
}
