% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotCoordinatesMarkers}
\alias{plotCoordinatesMarkers}
\title{Plot a marker for each cluster using parallel coordinates}
\usage{
plotCoordinatesMarkers(
  Celldata,
  condition.samples = c("condition", "timepoint"),
  samples = NULL,
  clusters = NULL,
  markers
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{condition.samples}{a character vector containing the variables to be studied for the samples. Possible values are: 'condition' or 'timepoint'}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{marlers}{a character vector containing the name of the markers to use}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualise the expression of a given marker in each cluster using parallel coordinates.
Each line in the representation corresponds to a biological sample for which marker/gene expressions are displayed.
}
