% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotDistogram}
\alias{plotDistogram}
\title{Plots of marker co-expression}
\usage{
plotDistogram(UMAPdata, clusters = NULL, show.on.device = TRUE)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{clusters}{a character vector containing the identifier of the cluster to use}

\item{show.on.device}{a numeric value containg xxx}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize xxx
}
