% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{plotHmAbundances}
\alias{plotHmAbundances}
\title{Plots an heatmap of cell cluster abundances}
\usage{
plotHmAbundances(
  Celldata,
  clusters = NULL,
  samples = NULL,
  saturation = 2.5,
  rescale = FALSE
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{saturation}{a numeric value providing the saturation threshold of cell cluster abundances}

\item{rescale}{a boolean specifying if cell cluster abundances must be centered and reduced}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the abundances of cell clusters using an heatmap representation.

In such heatmap each column corresponds a cell cluster and the row corresponds the different samples.
The heatmap can be restricted to specific cell clusters and samples.
The levels of abundance of each sample in each cluster is represented using a color gradient scale.
Abundance values can be centered and reduced.
}
