% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{plotHmExpressions}
\alias{plotHmExpressions}
\title{Plots an heatmap of cell marker expressions}
\usage{
plotHmExpressions(
  UMAPdata,
  markers = NULL,
  clusters = NULL,
  method.hclust = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  nb.cat = 5,
  seed = 42
)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{markers}{a character vector providing the marker names to use. By default, all markers are used}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{method.hclust}{a character value providing the agglomeration method to be use. Possible values are: 'ward.D', 'ward.D2', 'single', 'complete', 'average', 'mcquitty', 'median' or 'centroid' (please refer to the function 'hclust' of the 'stats' package)}

\item{nb.cat}{a numeric specifying the number of categories to use}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the cell marker expressions for selected markers and clusters

The mean of median marker expressions is computed for each cluster, and marker expressions displayed using a categorical heatmap (5 categories are defined by default)
The range expression of each cell marker is discretized into several categories between bounds of marker expressions
To hierarchical clustering, shown using dendrogramm, can be computed on both marker and cluster levels
}
