% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{plotHmStatistics}
\alias{plotHmStatistics}
\title{Plots an heatmap of a statistical analysis results}
\usage{
plotHmStatistics(
  Celldata,
  clusters = NULL,
  statistics = c("pvalue", "lfc", "effsize"),
  saturation = 3
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{statistics}{a character value providing the name of the statistic to display. Possible values are: 'pvalue' for p-value, 'lfc' for log2 fold change or 'eff' for effect size}

\item{saturation}{a numeric value providing the saturation value for statistics to display}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the results of differential cell clusters analysis

This representation displays statistical information for each cell cluster for a given comparison of samples
Different statistics can be visualized, such as the p-value, the log2(fold-change), and effect size
}
