% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotLDA}
\alias{plotLDA}
\title{Plots a LDA representation based cell cluster abundances}
\usage{
plotLDA(UMAPdata)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}
}
\value{
xx
}
\description{
This function aims to represent a Linear Discriminant Analysis representation based on cell cluster abundances.
}
