% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotMarkerDensity}
\alias{plotMarkerDensity}
\title{Plots the phenotype of identified cell clusters}
\usage{
plotMarkerDensity(
  Celldata,
  clusters,
  quant.low = 0.05,
  quant.high = 0.95,
  dip.th = 0.01
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifier of the cluster to use}

\item{quant.low}{a numeric value providing the value of first quantile in the color gradiant}

\item{quant.high}{a numeric value providing the value of last quantile in the color gradiant}

\item{dip.th}{a numeric value specifying the p-value threshold of the Hartigan's dip test to consider non-unimodal clusters}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the density of marker/gene expressions for a set of given clusters.
}
