% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotMDS}
\alias{plotMDS}
\title{Plots a MDS representation based on cell cluster abundances}
\usage{
plotMDS(
  UMAPdata,
  levels = c("clusters", "samples"),
  condition.samples = c("condition", "timepoint"),
  clusters = NULL,
  samples = NULL,
  plot.interactive = TRUE
)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{levels}{a character value containing the variable to be displayed. Possible values are: 'clusters' or 'samples'}

\item{condition.samples}{a character vector containing the variable to be studied for the samples. Possible values are: 'condition' or 'timepoint"}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{plot.interactive}{xxx}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the similarities between samples or clusters based on their abundances, using a Multidimensional Scaling representation.
Each dot represents a sample or a cluster and the distances between the dots are proportional to the Euclidean distance between these objects.
The representation can be restricted to specific cell clusters and samples. In addition, it is possible to choose the levels displayed, clusters or samples.
}
