% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotPCA}
\alias{plotPCA}
\title{Plots a PCA representation based cell cluster abundances}
\usage{
plotPCA(
  UMAPdata,
  levels = c("both", "clusters", "samples"),
  clusters = NULL,
  samples = NULL,
  components = c(1, 2),
  condition.samples = c("condition", "timepoint"),
  cor.radius.th = 0.6,
  cluster.label.size = 3,
  sample.label.size = 3,
  cluster.dot.size = 2,
  sample.dot.size = 3
)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{levels}{a character value containing the variable to be displayed. Possible values are: 'clusters' or 'samples'}

\item{clusters}{a character vector containing the identifier of the cluster to use. By default, all clusters are used}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{components}{a numeric vector providing the components to display}

\item{condition.samples}{a character vector containing xxx}

\item{cor.radius.th}{a numeric value containing}

\item{cluster.label.size}{a numeric value containing xxx}

\item{sample.label.size}{a numeric value containing xxx}

\item{cluster.dot.size}{a numeric value containing xxx}

\item{sample.dot.size}{a numeric value containing xxx}
}
\value{
a ggplot2 object
}
\description{
This function aims to represent a Principal Component Analysis representation based on cell cluster abundances.
In such representation, clusters or samples are positioned based on computed principal components. 
The representation can be displayed based on specific principal components. 
The representation can be restricted to specific cell clusters and samples. In addition, it is possible to choose the levels displayed, clusters or samples.
}
