% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotScatter}
\alias{plotScatter}
\title{Plots of scatter}
\usage{
plotScatter(UMAPdata, marker1, marker2, samples = NULL, clusters = NULL)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{marker1}{a character value containing xxx}

\item{marker2}{a character value containing xxx}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}
}
\value{
xx
}
\description{
This function aims to visualize xxx
}
