% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotSmallClusters}
\alias{plotSmallClusters}
\title{Internal - Plots a representation of QC for small clusters}
\usage{
plotSmallClusters(values.small)
}
\arguments{
\item{values.small}{a list providing the small cluster QC information. Such as data.frame containing the boolean values and the percentage computed.}
}
\value{
a ggplot2 object
}
\description{
This function is used internally to create a representation showing the fraction of clusters having a number associated cell lower than a specific threshold.
}
