% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plotUniformClusters}
\alias{plotUniformClusters}
\title{Internal - Plots a representation of QC for uniform phenotype cluster}
\usage{
plotUniformClusters(values_uniform)
}
\arguments{
\item{values_uniform}{a list providing the uniform cluster QC information. Such as data.frame containing the boolean values and the percentage computed.}
}
\value{
a ggplot2 object
}
\description{
This function is used internally to create a graphic representation.
}
