% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotVolcanoPlot}
\alias{plotVolcanoPlot}
\title{Plots of a volcano plot of statistical analysis}
\usage{
plotVolcanoPlot(UMAPdata, comparison, th.pv = 1.3, th.fc = 1.5)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{comparison}{a character value containing the comparison to study}

\item{th.pv}{a numeric value containing the p-value threshold to use}

\item{th.fc}{a numeric value containing the fold-change threshold to use}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the results of a differentially abundant a analysis using a Volcano plot.

In such representation, each in dot corresponds to a cell cluster and dots are positioned in two dimensional space where the X-axis represents the log2(fold-change) and the Y-axis represents the -log10 of the p-value.
Un horizontal line is displayed accordingly to the p-value threshold and to vertical lines are displayed accordingly to the fold-change threshold.
}
