% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCCorrelationManifold}
\alias{QCCorrelationManifold}
\title{QC for dimensionality reduction : Computes correlation between pairwise distances in the high-dimensional space and in the embedding}
\usage{
QCCorrelationManifold(
  Celldata,
  downsampling = 1000,
  method = c("pearson", "kendall", "spearman"),
  plot.device = TRUE
)
}
\arguments{
\item{Celldata}{a S4 object of class 'Celldata'}

\item{downsampling}{a numeric being the sample size used for computation}

\item{method}{a character being the method use to compute correlation. Possible values are : "pearson", "kendall", "spearman". Default value is "pearson"}

\item{plot.device}{a boolean value specifying if result representation must be displayed}
}
\value{
a list containing correlation coefficient and boxplot of pairwise distances
}
\description{
CPD quantifies preservation of the global, or macroscropic structure.
A performant RD should give a boxplot with linear relationships betweens pairwise distance in the different data spaces
}
