% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCKncManifold}
\alias{QCKncManifold}
\title{QC for dimensionality reduction : Computes proportion of nearest clusters preservation in reduced dimension in comparison to high dimension}
\usage{
QCKncManifold(Celldata, KNC = 5, downsampling = 50)
}
\arguments{
\item{Celldata}{a S4 object of class 'Celldata'}

\item{KNC}{a numeric being the number of nearest classes to compute}

\item{downsampling}{a numeric being the sample size used for computation}
}
\value{
a numeric being proportion
}
\description{
The fraction of k-nearest class (clusters) means in the original data that are preserved as k-nearest class means in the embedding.
it computes the class means only and averages across all classes. KNC quantifies preservation of the mesoscopic structure.
}
