% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCKnnManifold}
\alias{QCKnnManifold}
\title{QC for dimensionality reduction : Computes proportion of nearest neighbours preservation in reduced dimension in comparison to high dimension}
\usage{
QCKnnManifold(Celldata, KNN = 5, downsampling = 50)
}
\arguments{
\item{Celldata}{a S4 object of class 'Celldata'}

\item{KNN}{a numeric being the number of nearest neighbors to compute}

\item{downsampling}{a numeric being the sample size used for computation}
}
\value{
a numeric being proportion
}
\description{
The fraction of k-nearest neighbours in the original high-dimensional data that are preserved as k-nearest neighbours in the embedding
It compute the average across all n points. KNN quantifies preservation of the local, or microscopic structure.
}
