% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{QCMarkerNames}
\alias{QCMarkerNames}
\title{Verifies the consistency of the marker names within cell event files}
\usage{
QCMarkerNames(files)
}
\arguments{
\item{files}{a character vector specifying the path of the tab-separated or FCS files to check}
}
\value{
a data.frame containing the marker names and the associated number of cells for each sample (rownames = samples and colnames = markers)
}
\description{
This function aims to check the consistency of marker names across multiple tab-separated or FCS files.

Additionally, the number of cells associated to each sample is displayed.
}
