% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{QCMarkerRanges}
\alias{QCMarkerRanges}
\title{Verifies the consistency of marker expressions integrity within cell event files}
\usage{
QCMarkerRanges(files, probs = c(0.05, 0.95))
}
\arguments{
\item{files}{a character vector specifying the path of the FCS files to verified}

\item{probs}{a numerical vector providing the quantiles used to define marker expressions ranges}
}
\value{
a list containing two data.frame for the lower and upper marker expression ranges (rownames = samples and colnames = markers)
}
\description{
This function aims to check the consistency of marker expressions ranges across multiple tab-separated or FCS files.

The marker expressions ranges are calculated based on the user-defined quantiles.
}
