% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QCSmallClusters}
\alias{QCSmallClusters}
\title{Computes the percentage of cell clusters with low number of cells}
\usage{
QCSmallClusters(UMAPdata, th.size = 50, plot.device = TRUE)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{th.size}{a numeric value providing the minimum number of cells needed for a cluster to be considered a small cluster}

\item{plot.device}{a boolean value specifying a results representation must be displayed}
}
\value{
a numerical value corresponding to the percentage of cell cluster with low number of cells.
}
\description{
This function aims to compute and show cell clusters having a number of associated cells lower than a specific threshold.
}
