% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{renameMarkers}
\alias{renameMarkers}
\title{Renames markers within a UMAPdata object}
\usage{
renameMarkers(UMAPdata, marker.names)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{marker.names}{a character vector providing the new marker names to use}
}
\value{
a S4 object of class 'UMAPdata'
}
\description{
This function aims to rename cell markers stored within a UMAPdata object.

This function is interesting to remove the names of the fluorochromes or metals recorded during the acquisition process.
}
