% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{rescaleMarkerExpressions}
\alias{rescaleMarkerExpressions}
\title{Internal - Rescales marker expression by quantile}
\usage{
rescaleMarkerExpressions(exprs, quant.low = 0.05, quant.high = 0.95)
}
\arguments{
\item{exprs}{a data.frame containing the marker expressions for each cluster}

\item{quant.low}{a numeric value providing the value of the first quantile}

\item{quant.high}{a numeric value providing the value of the last quantile}
}
\value{
a data.frame containing quantile rescale marker expressions
}
\description{
This function is used internally to rescale the marker expression by the quantile method.
}
