% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{selectSamples}
\alias{selectSamples}
\title{Select samples based on metadata information}
\usage{
selectSamples(Celldata, individual = NULL, condition = NULL, timepoint = NULL)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{individual}{a character vector indicating the individual to select}

\item{condition}{a character vector indicating the biological condition to select}

\item{timepoint}{a character vector indicating the timepoint to select}
}
\value{
a character vector
}
\description{
This function aims to select biological samples of interest based on provided metadata.
}
