% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{show}
\alias{show}
\alias{show,Celldata-method}
\title{Prints information for a Celldata objects}
\usage{
\S4method{show}{Celldata}(object)
}
\arguments{
\item{object}{a Celldata object}
}
\value{
none
}
\description{
Shows a preview for a Celldata object
}
