% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Computes differential analysis statistics for cell clusters}
\usage{
computeStatistics(
  UMAPdata,
  condition,
  ref.condition,
  test.statistics = c("wilcoxon", "t-test"),
  paired = c("paired", "unpaired")
)
}
\arguments{
\item{UMAPdata}{a UMAPdata object}

\item{condition}{a character value providing the name of the condition to be compared}

\item{ref.condition}{a character value providing the name of reference condition}

\item{test.statistics}{a character value providing the type of statistical test to use. Possible values are: 'wilcoxon' or 't-test'}

\item{paired}{a boolean value indicating if a paired or unpaired comparison should be applied}
}
\value{
a S4 object of class 'UMAPdata'
}
\description{
This function aims to compute the statistics of Differentially Abundant Clusters 'DAC'. 

DAC correspond to cell clusters having abundances statistically different between two biological conditions.
The statistical test used for the comparisons can be defined by users. 
For each cluster, the p-value, log2 fold-change and effect size relative to the reference condition are computed.  
Statistical comparison can be performed in a paired and unpaired manner.
}
