% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{CreateMetaclusters}
\alias{CreateMetaclusters}
\title{Create metaclusters}
\usage{
CreateMetaclusters(Celldata, clusters, metaclusters)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{metaclusters}{a character value containing the name of the metacluster to create}
}
\value{
a Celldata object
}
\description{
This function aims to gathered mutliple cell cluster to a large cell cluster
}
