% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{createMetaclusters}
\alias{createMetaclusters}
\title{Create metaclusters}
\usage{
createMetaclusters(Celldata, clusters, metacluster.name)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifiers of the clusters to gather}

\item{metacluster.name}{a character value containing the name of the metacluster to create}
}
\value{
a Celldata object
}
\description{
This function aims to gathered multiple cell clusters to a large cell cluster
}
