% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\name{generateManifold}
\alias{generateManifold}
\title{Generates a manifold of cell events}
\usage{
generateManifold(
  Celldata,
  method = c("UMAP", "tSNE", "lvish"),
  markers = NULL,
  seed = 42,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{method}{a character value specifying the type of manifold to compute. Possible values are: 'UMAP' for Uniform Manifold Approximation and Projection, 'tSNE' for t-distributed Stochastic Neighbor Embedding, and 'lvish' for LargeVis}

\item{markers}{a character vector providing the cell markers to use for the manifold generation}

\item{seed}{a numeric value providing the random seed to use during stochastic operations}

\item{verbose}{a boolean value indicating if computational details must be displayed on the console}

\item{...}{other arguments passed on to manifold methods}
}
\value{
a S4 object of class 'Celldata'
}
\description{
This function aims to generate a manifold representation for cell events stored in a Celldata object.

This function allows the use UMAP, t-SNE or LargeVis dimension reduction techniques.
The whole set of cell markers or specific cell markers can be used during the dimensionality reduction process.
The random seed can also be defined to allow the reproducibility of generated results.
}
