% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importMTX}
\alias{importMTX}
\title{Imports of cell expression profiles from MTX files}
\usage{
importMTX(count, cells, features, meta, sample.col)
}
\arguments{
\item{count}{a character vector specifying the path of the count mtx file containing the count values}

\item{cells}{a character vector specifying the path of the mtx file containing the cell ids}

\item{features}{a character vector specifying the path of the mtx file containing the gene ids}

\item{meta}{a character specifying the name of the tsv metadata file to load}

\item{sample.col}{a character specifying the name of the colmun to use as biological sample in the metadata file}
}
\value{
a S4 object of class 'Celldata'
}
\description{
This function aims to import acquired cell events from single-cell transcriptomic profiling into a Celldata object.
}
