% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsampling.R
\name{performUpsampling}
\alias{performUpsampling}
\title{Performs the upsampling of downsampled events}
\usage{
performUpsampling(
  Celldata,
  files,
  transform = c("logicle", "arcsinh", "logarithmic", "none")
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{files}{a character vector providing the path of the tab-separated or FCS files}

\item{transform}{a character value containing the type of the transformation to apply. Possible values are: 'logicle', 'arcsinh', 'logarithmic' or 'none'}
}
\value{
a S4 object of class 'Celldata'
}
\description{
This function aims to perform the upsampling of downsampled events events based on an existing Celldata object and existing cell events stored in tab-separated or FCS files.

Importantly, the identification of cell clusters must have been performed prior to this operation.
}
