% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotClustersCounts}
\alias{plotClustersCounts}
\title{Plots the numbers of cells of each cluster}
\usage{
plotClustersCounts(
  Celldata,
  clusters = NULL,
  sort = TRUE,
  legend.max.samples = 10
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifiers of the clusters to use. By default, all clusters are used}

\item{sort}{a boolean value indicating if clusters must be sorted by the number associated cluster}

\item{legend.max.samples}{a numerical value specifying the maximal number of samples to display in the graphical legend}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the number of cells associated to each cluster.

This representation displays the clusters in the X-axis and the total number of associated cells in the Y-axis.
}
