% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotCoordinates}
\alias{plotCoordinates}
\title{Plots of phenotype of cell clusters using parallels coordinates}
\usage{
plotCoordinates(
  Celldata,
  condition.samples = c("condition", "timepoint"),
  samples = NULL,
  clusters
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{condition.samples}{a character vector containing the variables to be studied for the samples. Possible values are: 'condition' or 'timepoint'}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{clusters}{a character vector containing the identifiers of the clusters to use}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the characteristics of cell clusters using parallels coordinates.
Each line in the representation corresponds to a biological sample for which marker/gene expressions are displayed.
}
