% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotDistogram}
\alias{plotDistogram}
\title{Plots of a distogram of marker co-expression}
\usage{
plotDistogram(Celldata, clusters = NULL)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifier of the cluster to use}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize the pairwise co-expression between all markers using a distogram representation.
Each tile corresponds to the co-expression between two markers and is gradient-colored based on the Pearson correlation
}
\details{
The Pearson correlation is computed based on the marker expressions
}
