% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotManifold}
\alias{plotManifold}
\title{Plots a representation of a computed manifold}
\usage{
plotManifold(
  Celldata,
  markers = "density",
  samples = NULL,
  observation = c("individual", "condition", "timepoint"),
  scale = FALSE,
  quant.low = 0.05,
  quant.high = 0.95
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{markers}{a character value providing the name of the marker to use for the colouring. By default, cells are colored based on their local density}

\item{samples}{a character vector containing the names of biological samples to use}

\item{observation}{a character value containing the parameters to use}

\item{scale}{a boolean value specifing if expression calue must be rescaled}

\item{quant.low}{a numeric value providing the number of first quantile}

\item{quant.high}{a numeric value providing the number of last quantile}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize a computed manifold representation for given analysis.

This representation can be used on a Celldata object for which a manifold analysis has been performed.

If a cell clustering has been performed, then the clusters are delineated using concave hulls.
Additionnaly, the manifold can be colored based on the local cell density or marker expressions.
It is possible to centred ans reduce the values of expressions.
}
