% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotMarkerDensity}
\alias{plotMarkerDensity}
\title{Plots of phenotype of identified cell clusters}
\usage{
plotMarkerDensity(
  Celldata,
  clusters,
  quant.low = 0.05,
  quant.high = 0.95,
  dip.th = 0.01
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{clusters}{a character vector containing the identifier of the cluster to use}

\item{quant.low}{a numeric value providing the number of first quantile}

\item{quant.high}{a numeric value providing the number of last quantile}

\item{dip.th}{a numeric value specifing the p-value threshold of the Hartigan's dip test}
}
\value{
a ggplot2 object
}
\description{
This function aims to visualize xxx
}
