% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plotPCA}
\alias{plotPCA}
\title{Plots a PCA representation based cell cluster abundances}
\usage{
plotPCA(
  Celldata,
  levels = c("both", "clusters", "samples"),
  clusters = NULL,
  samples = NULL,
  components = c(1, 2),
  condition.samples = c("condition", "timepoint"),
  cor.radius.th = 0.6,
  plot.text = TRUE
)
}
\arguments{
\item{Celldata}{a Celldata object}

\item{levels}{a character value containing the variable to be displayed. Possible values are: 'both', 'clusters' or 'samples'}

\item{clusters}{a character vector containing the identifier of the cluster to use. By default, all clusters are used}

\item{samples}{a character vector containing the names of biological samples to use. By default, all samples are used}

\item{components}{a numeric vector providing the components to display}

\item{condition.samples}{a character vector containing the variable to be studied for the samples. Possible values are: 'condition' or 'timepoint"}

\item{cor.radius.th}{a numeric value specifing xxx}

\item{plot.text}{a boolean value specifing if adds text directly at the plot}
}
\value{
a ggplot2 object
}
\description{
This function aims to represent a Principal Component Analysis representation based on cell cluster abundances.
In such representation, clusters or samples are positioned based on computed principal components.
The representation can be displayed based on specific principal components.
The representation can be restricted to specific cell clusters and samples. In addition, it is possible to choose the levels displayed, clusters or samples.
}
