% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print,Celldata-method}
\title{Prints information for a given Celldata object}
\usage{
\S4method{print}{Celldata}(x)
}
\arguments{
\item{x}{a Celldata object}
}
\value{
none
}
\description{
Prints a preview for a Celldata object
}
