% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print,UMAPdata-method}
\title{Prints information for a given UMAPdata object}
\usage{
\S4method{print}{UMAPdata}(x)
}
\arguments{
\item{x}{a UMAPdata object}
}
\value{
none
}
\description{
Prints a preview for a UMAPdata object.
}
