% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{show}
\alias{show}
\alias{show,UMAPdata-method}
\title{Prints information for a UMAPdata objects}
\usage{
\S4method{show}{UMAPdata}(object)
}
\arguments{
\item{object}{a UMAPdata object}
}
\value{
none
}
\description{
Shows a preview for a UMAPdata object.
}
